made into component with props
